Refactor combat mechanics to use u16 for health and related attributes
- Changed health type from u32 to u16 in multiple structs and functions to optimize memory usage.
- Updated calculations and damage handling to accommodate the new health type.
- Introduced Randomness struct to handle randomness generation and improve critical hit calculations.
- Adjusted attack speeds and buffs to use u16 instead of u32 for consistency.
- Enhanced ability effects to support new buff types and improved packing/unpacking methods.
- Updated dependencies in Scarb.toml files to include necessary libraries for packing and randomness.

+++ packages/arcade/src/table.cairo
@@ -14,7 +14,7 @@ pub struct ArcadeAttempt {
     pub expiry: u64,
     pub abilities: Abilities,
     pub attacks: Span<felt252>,
-    pub health_regen: u32,
+    pub health_regen: u16,
     pub respawns: u32,
     pub stage: u32,
     pub phase: ArcadePhase,
